Tile Calculator written in angularJs
enter the dimensions of the room. Then enter the dimensions of the tile
you like and the calculator will tell you how many tile you need to
complete the job. You may also enter in the waste percentage of tiles
you would like as well as the price per tile and the calculator will
give you the number of tiles need with the waste percentage and the
total cost of the project.